Refactor out to a concern.
