+++ services/QuillLMS/engines/evidence/lib/evidence/evidence/open_ai/completion.rb
@@ -4,6 +4,7 @@ module Evidence
   module OpenAI
     class Completion < Evidence::ApplicationService
       include Evidence::OpenAI::APIConcern
+      include Evidence::OpenAI::SentenceResultsConcern
 
       ENDPOINT = '/completions'
